Create README.md
I just received my comma 3x yesterday and was able to install it in my car. Unfortunately, when I upload the application, I receive the error "Car Unsuported". I have a 2020 Hyundai Kona Unlimited with all safety features, and it is listed as a compatible vehicle in parenthesis. Could you please advise me on how to solve this problem? Thanks,

Dennis